requirements.txt
Django>=5.0,<6.0
django-ninja>=1.1,<2.0
django-cors-headers>=4.3,<5.0
PyJWT>=2.8,<3.0
Pillow>=10.0,<11.0
python-decouple>=3.8,<4.0
psycopg2-binary>=2.9,<3.0
python-dateutil>=2.8,<3.0
